+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Change Log
 
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
+
+[0.17.2]
+********
+* Updated state, health state and load dish config result event receiver with queue mechanism.
+* Removed unused attributes from mid and low.
+  
 [0.17.1]
 ********
 * Fix telescope ON issue.